ignore state loss during fragment transaction
We don't need to worry about state loss here. If the fragment state is already
saved, the loader will re-fire when the fragment resumes and we'll proceed to
mailbox selection / account list as expected.

Also; only add the account selection fragment to the back stack if there is
more than one account.

bug 4558901

Change-Id: I38290fa378ac80ef5a5c8cf25c38b6e820715366