NFC zero-click sharing for Email.
This shares the e-mail address of the currently shown account
over P2P NFC links. Note that the currently set account must
be updated actively, because the callback to create the message
comes from one of the NFC threads, which don't have permissions
to retrieve the account and the correspondig e-mail address.

Change-Id: I12e48a3787d469179e7ca074429106ed24e3a212