Fix search styling.
This eliminates the excessive padding on the left and makes the hint
text behave according to the new spec (since it uses the new action view
expanded mode).

A later CL will actually change the item to use the action view
facilities of the framework (so we don't have to manage our own custom
views).

Bug: 5109953
Change-Id: Ia3e4a1e64e65b3b8bb3e708b31636f4186acfd46