Remove needless clearView()
- this is leftover from a time where we re-used a single WebView for
multiple messages. The current architecture does not rely on that, so
the clearView() is actually not useful. Worse - it's causing issues
because the operation in the WebView is asynchronous, so a clearView and
a load of the content in quick succession causes a race condition,
causing items not to load

Bug: 5205296
Change-Id: I5ee6697bdf53a033ba1c7f21170820259bf830d1