Fix widget updates
Fix widgets that stop updating.  If the Launcher got killed, the Email
widget service gets unbound.  When this happens the cursor loader for the
widget is stopped.  Since widget is relying on the loader to update the data,
when the loader is stopped, the widget would never update.

Now when email recognizes a change, it will send a broadcast intent, which
will cause the widget service to be started, if it isn't.

Bug: 5811810
Change-Id: Ia840e58f10e780b94440119662c2e48e7785c507