Turkish characters show problem in addressList and subject
Set Locale.US as parameter for toLowerCase() to make sure
that the string is changed to lowercase correctly, e.g.
'ISO-8859-9' to 'iso-8859-9'. The email code supports the
decoding of charset 'iso-8859-9', so that the encoded address's
name and subject, will be decoded and display the correct
content in the phone.

Change-Id: Id1a7cda0bbf83a7b584ab12336ff66195835a2fc