Fix crash when the Email process is killed
If the Email process is killed during setup, the Account information
wouldn't be set.  Specify this information in the intent as the intent will be delivered
again when the activity is restored

This is a short term fix. Ideally we would switch all of the intents for the setup
activities to use these extras to pass this information, or we would save the state of
SetupData in onSaveInstanceState

Bug: 7163201

Change-Id: Ic1758eb031cd113089073d03f3c2571e37a98b23