Prompt for auto-advance option on first action
If we're on a tablet, we will pop up a dialog that asks the user what
they want to do. When they select an option, we save it, and take
the original action, and auto-advance to their preference.

Depends on I7b11bd0d17e6cc184b67c8c943531146fc964b49

Bug: 7328118
Change-Id: Ied57eee4d90ad1d61316032d0796f9205fdd1962