Add user assignable category for mail foldders if it accepts moved messages
Email requires a user assignable category in the case
the user is moving labels/folders from a category the
user should not be allowed to view (i.e. sys labels).

Change added flag under function where if it allows
moved messages, it will allow the user to assign
there as well.

Bug: 8157593

Change-Id: Icc8a5d4587ef9180d32207613da4f31ddc8854f1