Fix NPE when adding account with security policy
When adding an Exchange account with a security policy, Email
generates a notification to let the user enable the policy. The
code that generates the notification attemptes to load the folder
as part of the base notification.  Before the account policy has been
configured, it isn't possible to load the folder.  We need to handle
a null folder when generating the notification

Also, imporoved the logging when an exception is thrown from the
PolicyService

Bug: 8341882
Change-Id: I33ff57adc3c130e8318910e352146b8ea56f9a2c