Adding drag-state drawable to folder item icons.
Fixes b/8986628. Previously we had two states for
icons in the folder list (activated and default).
We now have a third state for when the app is in
drag mode and the folder is disabled which uses
a lighter grey icon. To make this work, we use
a custom drawable state (state_drag_mode) for
when drag is entered. Another CL will use this
new drag state for changing folder text color as well.

Change-Id: I4c752ac7b996fd026e706f8e592ec57326ac61bf
(cherry picked from commit 0670cddaf92475ff30c7bd05af337ffa12e9ca95)