Update version number
This also changes the version number format. Instead
of being a single increment each time, we're actually
encoding the date in the version number.

The format is now vvMMDDc

where vv is major version number
MM is the ordinal month number. This starts at 0 for each major release and
   counts up for each month
DD is the day of the month
c is a code which increments per version, in case we have multiple releases
   on the same date. This should start at zero.

Change-Id: I006745754614e9c72261035748a90a38871b67e6