Move Help & feedback to Nav drawer.
b/11061485

HELP and SEND FEEDBACK are now the two bottom-most drawer
items in the Navigation drawer when it is present or in
the overflow menu when it is not present.

SEND FEEDBACK still behaves as always.

HELP has been redesigned so that it displays the help
content in a local webview rather than using an external
browser.

Change-Id: I01003f3cd6f8dedbcfb95ae87872968dfc6be48b