b/13659097. Disable EAS Policy max attachment size from EmailProvider.
Disable the code that translates the max attachment size column of
our db to the EmailProvider max attachment size column. They mean
different things as our max attachment size column refers to the max
size of any file that we can download where the EmailProvider's version
represents the max attachment size of a file that we can upload.

Change-Id: I93202766d7b1782c902f17ce20b2823aaf647ec4