b/13659368. Ensure properly download status for attachments on failures.
The reason why the UI is wedged is becasue it doesn't
see a state transition in the status. So now we detect the situation
where the user is trying to download an ineligible file and force
the status to indicate that it was a failure to dismiss the UI and
to show the correct error message.

Change-Id: I9264966830a317724bf2fe469ae570860ba7c2a9