b/13734162. Make sure that SmartForward is enabled on EAS accounts.
This is an Algol only fix that needs to be reconsidered for L.
The fix comes in 2 parts. The first addresses existing accounts.
The second addresses new accounts. First we check that the protocol
is EAS and then we check the protocol version to be greater than
12.0 before setting the flags. In the future, the protocol specific
flags should be set in the protocol specific components.

Change-Id: I00a6fa3337d5230801f6de8876332c66deeae2c4